requirements.txt
bcrypt==3.2.0
cffi==1.14.6
cryptography==3.4.7
et-xmlfile==1.1.0
greenlet==1.1.0
importlib-metadata==4.6.1
numpy==1.21.0
openpyxl==3.0.7
pandas==1.3.0
Pillow==8.3.1
pycparser==2.20
PyNaCl==1.4.0
pyodbc==4.0.31
python-dateutil==2.8.1
python-dotenv==0.18.0
python-magic==0.4.24
python-magic-bin==0.4.14
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.20
typing-extensions==3.10.0.0
zipp==3.5.0
